refactor(cli): extract telemetry collection from monitor.rs
## Summary
Extracts /proc/meminfo parsing, vmstat collection, and GPU utilization polling into `telemetry_collector.rs` from the 82KB `monitor.rs` god-file. This improves modularity and organization within `ramshared-cli`.

## Commits
- 084ab222cd58b08be5a2a4a5f2b3d337130f1613: refactor(cli): extract telemetry collection from monitor.rs

## Validation
- `cargo check --all-targets` and `cargo clippy` pass cleanly with 0 warnings.
- `cargo test -p ramshared-cli` passes completely with zero regressions.

## Rollback trigger
Rollback if telemetry collection drops fields, fails compilation, or fails code review.
